# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UptimePulse API
  version: 0.0.1
  description: |
    API REST de UptimePulse (monitorización de uptime). Se sirve en `/docs`
    (Swagger UI) desde la propia API.

    **Autenticación.** Dos formas, ambas por `Authorization: Bearer …`:
    - un *access token* JWT de 5 min (de `/auth/register`, `/auth/login` o
      `/auth/refresh`; el refresh token viaja en una cookie httpOnly y caduca
      a los `SESSION_IDLE_TIMEOUT_MINUTES` de inactividad, 15 por defecto,
      renovándose en cada refresh);
    - una **API key** `up_…` creada por un admin en
      `/organizations/{id}/api-keys`. Actúa en nombre de la organización con
      el rol equivalente a sus scopes (`read` → readonly, `write` → editor) y
      nunca puede administrar la organización (`/me`, equipo, claves).

    **Organización activa.** Las rutas de datos actúan sobre la organización
    indicada en `X-Organization-Id`; sin cabecera, sobre la personal del
    usuario. Con API key la fija la clave y la cabecera se ignora.

    **Errores.** Siempre `{ "error": "<mensaje en español>" }`; los 400 de
    validación pueden llevar el detalle de zod en `error`. Cada respuesta
    lleva `X-Request-Id`.

    **Rate limits.** 300 peticiones/min por IP o API key en toda la API (429);
    login y registro 10/min; `/public/status`, `/invitations/{token}`
    y `/auth/username-available` 60/min.
servers:
  - url: /api
    description: Self-hosted (detrás del nginx del frontend, mismo origen)
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: JWT de acceso o API key `up_…`.
  parameters:
    organizationHeader:
      name: X-Organization-Id
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Organización activa. Sin ella, la personal del usuario.
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          description: Mensaje legible, o el detalle de validación de zod.
          oneOf: [{ type: string }, { type: object }]
...
    MonitorInput:
      type: object
      required: [name, type, target, intervalSeconds, timeoutMs]
      properties:
        name: { type: string, maxLength: 200 }
        type: { $ref: "#/components/schemas/MonitorType" }
        target:
          type: string
          description: |
            `http`: URL completa. `tcp`: `host:puerto`. `ping`: host o IP sin
            esquema ni puerto. Los destinos privados/internos (127.0.0.0/8,
            10/8, 172.16/12, 192.168/16, 169.254/16, ::1, fc00::/7…) se
            rechazan con 422, también si el DNS resuelve a uno.
        method: { type: string, enum: [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS], default: GET, description: Solo http. }
        headers: { type: object, additionalProperties: { type: string }, description: Solo http. }
        body: { type: string, description: Solo http. }
        expectedStatus: { type: integer, minimum: 100, maximum: 599, description: "Solo http. Sin él, cualquier status < 400 es operativo." }
        intervalSeconds: { type: integer, minimum: 30, maximum: 86400 }
        timeoutMs: { type: integer, minimum: 1000, maximum: 60000 }
        tags: { type: array, items: { type: string } }
    Monitor:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        name: { type: string }
        type: { $ref: "#/components/schemas/MonitorType" }
        target: { type: string }
        method: { type: string, nullable: true }
        headers: { type: object, nullable: true }
        body: { type: string, nullable: true }
        expectedStatus: { type: integer, nullable: true }
        intervalSeconds: { type: integer }
        timeoutMs: { type: integer }
        isPaused: { type: boolean }
        tags: { type: array, items: { type: string } }
        consolidatedStatus:
          type: string
          enum: [up, degraded, down]
          nullable: true
          description: Estado consolidado multi-región (null hasta el primer check).
        sslExpiresAt: { type: string, format: date-time, nullable: true }
        lastCheck: { $ref: "#/components/schemas/Check" }
        createdAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Sin token, token inválido/caducado o API key revocada.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Sin rol suficiente, organización ajena, o acción de persona intentada con API key.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    BadRequest:
      description: Validación.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
paths:
...
  /monitors:
    get:
      tags: [Monitores]
      summary: Monitores de la organización activa, con su último check
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Monitor" } } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Monitores]
      summary: Crear (editor). Programa el check periódico y uno inmediato.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/MonitorInput" } } }
      responses:
        "201": { description: "Creado.", content: { application/json: { schema: { $ref: "#/components/schemas/Monitor" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: Target privado/interno o que no resuelve.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /monitors/summary:
    get:
      tags: [Monitores]
      summary: Resumen del dashboard (contadores por estado y sparklines 24h)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Resumen." }
  /monitors/{id}:
    get:
      tags: [Monitores]
      summary: Detalle con el último check por región
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: "Detalle.", content: { application/json: { schema: { $ref: "#/components/schemas/MonitorDetail" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Monitores]
      summary: Editar (editor). Reprograma el check si cambia el intervalo.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      requestBody:
        content: { application/json: { schema: { $ref: "#/components/schemas/MonitorInput" } } }
      responses:
        "200": { description: "Actualizado.", content: { application/json: { schema: { $ref: "#/components/schemas/Monitor" } } } }
        "422": { description: "Target privado/interno o que no resuelve.", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    delete:
      tags: [Monitores]
      summary: Borrar (editor). Quita sus job schedulers en todas las regiones.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "204": { description: "Borrado." }
        "404": { $ref: "#/components/responses/NotFound" }
  /monitors/{id}/pause:
    post:
      tags: [Monitores]
      summary: Pausar (editor)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: "Pausado.", content: { application/json: { schema: { $ref: "#/components/schemas/Monitor" } } } }
  /monitors/{id}/resume:
    post:
      tags: [Monitores]
      summary: Reanudar (editor)
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: "Reanudado.", content: { application/json: { schema: { $ref: "#/components/schemas/Monitor" } } } }
  /monitors/{id}/checks:
    get:
      tags: [Monitores]
      summary: Últimos checks (todas las regiones)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200": { description: "Más recientes primero.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Check" } } } } }
...
  /monitors/{id}/timeseries:
    get:
      tags: [Monitores]
      summary: Serie por hora (para gráficos) en un rango
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/range" }]
      responses:
        "200": { description: "Puntos.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/TimeseriesPoint" } } } } }
  /monitors/{id}/incidents:
    get:
      tags: [Monitores]
      summary: Incidentes del rango
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }, { $ref: "#/components/parameters/range" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Incident" } } } } }
  /monitors/{id}/maintenance-windows:
    get:
      tags: [Monitores]
      summary: Ventanas de mantenimiento
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/MaintenanceWindow" } } } } }
    post:
      tags: [Monitores]
      summary: Crear ventana (editor). Dentro de ella no se abren incidentes ni se alerta.
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/MaintenanceWindowInput" } } }
      responses:
        "201": { description: "Creada.", content: { application/json: { schema: { $ref: "#/components/schemas/MaintenanceWindow" } } } }
  /monitors/{id}/maintenance-windows/{windowId}:
    delete:
      tags: [Monitores]
      summary: Borrar ventana (editor)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: windowId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: "Borrada." }
  /monitors/{id}/notification-channels:
    get:
      tags: [Monitores]
      summary: Canales asociados al monitor
      security: [{ bearer: [] }]
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "200": { description: "Lista.", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Channel" } } } } }
  /monitors/{id}/notification-channels/{channelId}:
    post:
      tags: [Monitores]
      summary: Asociar canal (editor)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: channelId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "201": { description: "Asociado." }
    delete:
      tags: [Monitores]
      summary: Desasociar canal (editor)
      security: [{ bearer: [] }]
      parameters:
        - { $ref: "#/components/parameters/id" }
        - { name: channelId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: "Desasociado." }
...